Updates seed to generate product categories
