+++ .env-example
@@ -2,9 +2,10 @@
 
 DATABASE_URL=postgres://postgres:postgres@localhost:5432/shopping-cart-dev
 
-# Seed product count defaults to 100, useful to specifying how many seed records to create
+# Seed data count defaults, useful to specifying how many seed records to create
 
 SEED_PRODUCT_COUNT=100
+SEED_CATEGORY_COUNT=10
 
 # Port to listen on for http
